+++ sdks/python/apache_beam/transforms/userstate_test.py
@@ -22,17 +22,25 @@ import unittest
 
 import mock
 
+import apache_beam as beam
 from apache_beam.coders import BytesCoder
+from apache_beam.coders import IterableCoder
 from apache_beam.coders import VarIntCoder
 from apache_beam.runners.common import DoFnSignature
+from apache_beam.testing.test_pipeline import TestPipeline
+from apache_beam.testing.test_stream import TestStream
+from apache_beam.transforms import userstate
+from apache_beam.transforms.combiners import ToListCombineFn
 from apache_beam.transforms.combiners import TopCombineFn
 from apache_beam.transforms.core import DoFn
 from apache_beam.transforms.timeutil import TimeDomain
 from apache_beam.transforms.userstate import BagStateSpec
 from apache_beam.transforms.userstate import CombiningValueStateSpec
 from apache_beam.transforms.userstate import TimerSpec
-from apache_beam.transforms.userstate import UserStateUtils
+from apache_beam.transforms.userstate import get_dofn_specs
+from apache_beam.transforms.userstate import is_stateful_dofn
 from apache_beam.transforms.userstate import on_timer
+from apache_beam.transforms.userstate import validate_stateful_dofn
 
 
 class TestStatefulDoFn(DoFn):
@@ -83,11 +91,11 @@ class InterfaceTest(unittest.TestCase):
     return DoFnSignature(dofn)
 
   @mock.patch(
-      'apache_beam.transforms.userstate.UserStateUtils.validate_stateful_dofn')
+      'apache_beam.transforms.userstate.validate_stateful_dofn')
   def test_validate_dofn(self, unused_mock):
     dofn = TestStatefulDoFn()
     self._validate_dofn(dofn)
-    UserStateUtils.validate_stateful_dofn.assert_called_with(dofn)
+    userstate.validate_stateful_dofn.assert_called_with(dofn)
 
   def test_spec_construction(self):
     BagStateSpec('statename', VarIntCoder())
@@ -117,8 +125,8 @@ class InterfaceTest(unittest.TestCase):
       DoFn.TimerParam(BagStateSpec('elements', BytesCoder()))
 
   def test_stateful_dofn_detection(self):
-    self.assertFalse(UserStateUtils.is_stateful_dofn(DoFn()))
-    self.assertTrue(UserStateUtils.is_stateful_dofn(TestStatefulDoFn()))
+    self.assertFalse(is_stateful_dofn(DoFn()))
+    self.assertTrue(is_stateful_dofn(TestStatefulDoFn()))
 
   def test_good_signatures(self):
     class BasicStatefulDoFn(DoFn):
@@ -133,14 +141,14 @@ class InterfaceTest(unittest.TestCase):
       def expiry_callback(self, element, timer=DoFn.TimerParam(EXPIRY_TIMER)):
         yield element
 
-    # Validate UserStateUtils.get_dofn_specs() and timer callbacks in
+    # Validate get_dofn_specs() and timer callbacks in
     # DoFnSignature.
     stateful_dofn = BasicStatefulDoFn()
     signature = self._validate_dofn(stateful_dofn)
     expected_specs = (set([BasicStatefulDoFn.BUFFER_STATE]),
                       set([BasicStatefulDoFn.EXPIRY_TIMER]))
     self.assertEqual(expected_specs,
-                     UserStateUtils.get_dofn_specs(stateful_dofn))
+                     get_dofn_specs(stateful_dofn))
     self.assertEqual(
         stateful_dofn.expiry_callback,
         signature.timer_methods[BasicStatefulDoFn.EXPIRY_TIMER].method_value)
@@ -153,7 +161,7 @@ class InterfaceTest(unittest.TestCase):
                            TestStatefulDoFn.EXPIRY_TIMER_2,
                            TestStatefulDoFn.EXPIRY_TIMER_3]))
     self.assertEqual(expected_specs,
-                     UserStateUtils.get_dofn_specs(stateful_dofn))
+                     get_dofn_specs(stateful_dofn))
     self.assertEqual(
         stateful_dofn.on_expiry_1,
         signature.timer_methods[TestStatefulDoFn.EXPIRY_TIMER_1].method_value)
@@ -270,7 +278,7 @@ class InterfaceTest(unittest.TestCase):
         (r'The on_timer callback for TimerSpec\(expiry1\) is not the '
          r'specified .on_expiry_1 method for DoFn '
          r'StatefulDoFnWithTimerWithTypo2 \(perhaps it was overwritten\?\).')):
-      UserStateUtils.validate_stateful_dofn(dofn)
+      validate_stateful_dofn(dofn)
 
     # (2) Here, the user forgot to add an on_timer decorator for 'expiry2'
     class StatefulDoFnWithTimerWithTypo3(DoFn):
@@ -299,7 +307,217 @@ class InterfaceTest(unittest.TestCase):
         ValueError,
         (r'DoFn StatefulDoFnWithTimerWithTypo3 has a TimerSpec without an '
          r'associated on_timer callback: TimerSpec\(expiry2\).')):
-      UserStateUtils.validate_stateful_dofn(dofn)
+      validate_stateful_dofn(dofn)
+
+
+class StatefulDoFnOnDirectRunnerTest(unittest.TestCase):
+  # pylint: disable=expression-not-assigned
+  all_records = None
+
+  def setUp(self):
+    # Use state on the TestCase class, since other references would be pickled
+    # into a closure and not have the desired side effects.
+    #
+    # TODO(BEAM-5295): Use assert_that after it works for the cases here in
+    # streaming mode.
+    StatefulDoFnOnDirectRunnerTest.all_records = []
+
+  def record_dofn(self):
+    class RecordDoFn(DoFn):
+      def process(self, element):
+        StatefulDoFnOnDirectRunnerTest.all_records.append(element)
+
+    return RecordDoFn()
+
+  def test_simple_stateful_dofn(self):
+    class SimpleTestStatefulDoFn(DoFn):
+      BUFFER_STATE = BagStateSpec('buffer', BytesCoder())
+      EXPIRY_TIMER = TimerSpec('expiry', TimeDomain.WATERMARK)
+
+      def process(self, element, buffer=DoFn.StateParam(BUFFER_STATE),
+                  timer1=DoFn.TimerParam(EXPIRY_TIMER)):
+        unused_key, value = element
+        buffer.add('A' + str(value))
+        timer1.set(20)
+
+      @on_timer(EXPIRY_TIMER)
+      def expiry_callback(self, buffer=DoFn.StateParam(BUFFER_STATE),
+                          timer=DoFn.TimerParam(EXPIRY_TIMER)):
+        yield ''.join(sorted(buffer.read()))
+
+    with TestPipeline() as p:
+      test_stream = (TestStream()
+                     .advance_watermark_to(10)
+                     .add_elements([1, 2])
+                     .add_elements([3])
+                     .advance_watermark_to(25)
+                     .add_elements([4]))
+      (p
+       | test_stream
+       | beam.Map(lambda x: ('mykey', x))
+       | beam.ParDo(SimpleTestStatefulDoFn())
+       | beam.ParDo(self.record_dofn()))
+
+    # Two firings should occur: once after element 3 since the timer should
+    # fire after the watermark passes time 20, and another time after element
+    # 4, since the timer issued at that point should fire immediately.
+    self.assertEqual(
+        ['A1A2A3', 'A1A2A3A4'],
+        StatefulDoFnOnDirectRunnerTest.all_records)
+
+  def test_simple_stateful_dofn_combining(self):
+    class SimpleTestStatefulDoFn(DoFn):
+      BUFFER_STATE = CombiningValueStateSpec(
+          'buffer',
+          IterableCoder(VarIntCoder()), ToListCombineFn())
+      EXPIRY_TIMER = TimerSpec('expiry1', TimeDomain.WATERMARK)
+
+      def process(self, element, buffer=DoFn.StateParam(BUFFER_STATE),
+                  timer1=DoFn.TimerParam(EXPIRY_TIMER)):
+        unused_key, value = element
+        buffer.add(value)
+        timer1.set(20)
+
+      @on_timer(EXPIRY_TIMER)
+      def expiry_callback(self, buffer=DoFn.StateParam(BUFFER_STATE),
+                          timer=DoFn.TimerParam(EXPIRY_TIMER)):
+        yield ''.join(str(x) for x in sorted(buffer.read()))
+
+    with TestPipeline() as p:
+      test_stream = (TestStream()
+                     .advance_watermark_to(10)
+                     .add_elements([1, 2])
+                     .add_elements([3])
+                     .advance_watermark_to(25)
+                     .add_elements([4]))
+      (p
+       | test_stream
+       | beam.Map(lambda x: ('mykey', x))
+       | beam.ParDo(SimpleTestStatefulDoFn())
+       | beam.ParDo(self.record_dofn()))
+
+    self.assertEqual(
+        ['123', '1234'],
+        StatefulDoFnOnDirectRunnerTest.all_records)
+
+  def test_timer_output_timestamp(self):
+    class TimerEmittingStatefulDoFn(DoFn):
+      EMIT_TIMER_1 = TimerSpec('emit1', TimeDomain.WATERMARK)
+      EMIT_TIMER_2 = TimerSpec('emit2', TimeDomain.WATERMARK)
+      EMIT_TIMER_3 = TimerSpec('emit3', TimeDomain.WATERMARK)
+
+      def process(self, element,
+                  timer1=DoFn.TimerParam(EMIT_TIMER_1),
+                  timer2=DoFn.TimerParam(EMIT_TIMER_2),
+                  timer3=DoFn.TimerParam(EMIT_TIMER_3)):
+        timer1.set(10)
+        timer2.set(20)
+        timer3.set(30)
+
+      @on_timer(EMIT_TIMER_1)
+      def emit_callback_1(self):
+        yield 'timer1'
+
+      @on_timer(EMIT_TIMER_2)
+      def emit_callback_2(self):
+        yield 'timer2'
+
+      @on_timer(EMIT_TIMER_3)
+      def emit_callback_3(self):
+        yield 'timer3'
+
+    class TimestampReifyingDoFn(DoFn):
+      def process(self, element, ts=DoFn.TimestampParam):
+        yield (element, int(ts))
+
+    with TestPipeline() as p:
+      test_stream = (TestStream()
+                     .advance_watermark_to(10)
+                     .add_elements([1]))
+      (p
+       | test_stream
+       | beam.Map(lambda x: ('mykey', x))
+       | beam.ParDo(TimerEmittingStatefulDoFn())
+       | beam.ParDo(TimestampReifyingDoFn())
+       | beam.ParDo(self.record_dofn()))
+
+    self.assertEqual(
+        [('timer1', 10), ('timer2', 20), ('timer3', 30)],
+        sorted(StatefulDoFnOnDirectRunnerTest.all_records))
+
+  def test_index_assignment(self):
+    class IndexAssigningStatefulDoFn(DoFn):
+      INDEX_STATE = BagStateSpec('index', VarIntCoder())
+
+      def process(self, element, state=DoFn.StateParam(INDEX_STATE)):
+        unused_key, value = element
+        next_index, = list(state.read()) or [0]
+        yield (value, next_index)
+        state.clear()
+        state.add(next_index + 1)
+
+    with TestPipeline() as p:
+      test_stream = (TestStream()
+                     .advance_watermark_to(10)
+                     .add_elements(['A', 'B'])
+                     .add_elements(['C'])
+                     .advance_watermark_to(25)
+                     .add_elements(['D']))
+      (p
+       | test_stream
+       | beam.Map(lambda x: ('mykey', x))
+       | beam.ParDo(IndexAssigningStatefulDoFn())
+       | beam.ParDo(self.record_dofn()))
+
+    self.assertEqual(
+        [('A', 0), ('B', 1), ('C', 2), ('D', 3)],
+        StatefulDoFnOnDirectRunnerTest.all_records)
+
+  def test_hash_join(self):
+    class HashJoinStatefulDoFn(DoFn):
+      BUFFER_STATE = BagStateSpec('buffer', BytesCoder())
+      UNMATCHED_TIMER = TimerSpec('unmatched', TimeDomain.WATERMARK)
+
+      def process(self, element, state=DoFn.StateParam(BUFFER_STATE),
+                  timer=DoFn.TimerParam(UNMATCHED_TIMER)):
+        key, value = element
+        existing_values = list(state.read())
+        if not existing_values:
+          state.add(value)
+          timer.set(100)
+        else:
+          yield 'Record<%s,%s,%s>' % (key, existing_values[0], value)
+          state.clear()
+          timer.clear()
+
+      @on_timer(UNMATCHED_TIMER)
+      def expiry_callback(self, state=DoFn.StateParam(BUFFER_STATE)):
+        buffered = list(state.read())
+        assert len(buffered) == 1, buffered
+        state.clear()
+        yield 'Unmatched<%s>' % (buffered[0],)
+
+    with TestPipeline() as p:
+      test_stream = (TestStream()
+                     .advance_watermark_to(10)
+                     .add_elements([('A', 'a'), ('B', 'b')])
+                     .add_elements([('A', 'aa'), ('C', 'c')])
+                     .advance_watermark_to(25)
+                     .add_elements([('A', 'aaa'), ('B', 'bb')])
+                     .add_elements([('D', 'd'), ('D', 'dd'), ('D', 'ddd'),
+                                    ('D', 'dddd')])
+                     .advance_watermark_to(125)
+                     .add_elements([('C', 'cc')]))
+      (p
+       | test_stream
+       | beam.ParDo(HashJoinStatefulDoFn())
+       | beam.ParDo(self.record_dofn()))
+
+    self.assertEqual(
+        ['Record<A,a,aa>', 'Record<B,b,bb>', 'Record<D,d,dd>',
+         'Record<D,ddd,dddd>', 'Unmatched<aaa>', 'Unmatched<c>',
+         'Unmatched<cc>'],
+        sorted(StatefulDoFnOnDirectRunnerTest.all_records))
 
 
 if __name__ == '__main__':
